+++ app/(web)/(base page)/(learn)/electives/components/Electives/index.tsx
@@ -9,7 +9,7 @@ import { ElectiveCourseType } from '@/service/webApi/elective/type';
 import CourseFilterListSearch from '../CourseFilterListSearch';
 import CourseFilterListDefault from '../CourseFilterListDefault';
 import { useRequest } from 'ahooks';
-import { errorMessage } from '@/helper/utils';
+import { errorMessage } from '@/helper/ui';
 import { Metadata } from 'next';
 
 export const metadata: Metadata = {
