+++ app/mobile/(learn)/electives/components/MobCourseFilterListSearch/index.tsx
@@ -6,7 +6,7 @@ import {
   courseDefaultSort as sort
 } from '@/components/Web/Business/CourseFilterList/constant';
 import { FilterParamsType } from '@/components/Web/Business/CourseFilterList/type';
-import { errorMessage } from '@/helper/utils';
+import { errorMessage } from '@/helper/ui';
 import webApi from '@/service';
 import { ElectiveCourseType } from '@/service/webApi/elective/type';
 import { useRequest } from 'ahooks';
