+++ app/mobile/(learn)/electives/components/index.tsx
@@ -7,7 +7,7 @@ import { ElectiveCourseType } from '@/service/webApi/elective/type';
 // import CourseFilterListSearch from '../CourseFilterListSearch';
 // import CourseFilterListDefault from '../CourseFilterListDefault';
 import { useRequest } from 'ahooks';
-import { errorMessage } from '@/helper/utils';
+import { errorMessage } from '@/helper/ui';
 import { Metadata } from 'next';
 import MobElectiveCard from '@/components/Mobile/MobElectiveCard';
 import MobViewMoreList from '@/components/Mobile/MobViewMoreList';
