+++ app/mobile/(learn)/practices/components/index.tsx
@@ -6,7 +6,7 @@ import webApi from '@/service';
 // import CourseFilterListSearch from '../CourseFilterListSearch';
 // import CourseFilterListDefault from '../CourseFilterListDefault';
 import { useRequest } from 'ahooks';
-import { errorMessage } from '@/helper/utils';
+import { errorMessage } from '@/helper/ui';
 import { Metadata } from 'next';
 import MobViewMoreList from '@/components/Mobile/MobViewMoreList';
 import MobCourseFilterListDefault from './MobCourseFilterListDefault';
