+++ app/mobile/(lesson page)/mini/[courseId]/learn/[lessonId]/MiniCoursePage/hooks/useGetElectives.ts
@@ -1,4 +1,4 @@
-import { errorMessage } from '@/helper/utils';
+import { errorMessage } from '@/helper/ui';
 import webApi from '@/service';
 import {
   ElectiveLessonType,
