+++ app/mobile/(profile)/user/profile/components/PersonalEdit/BackgroundImage/index.tsx
@@ -1,7 +1,8 @@
 import HoverIcon from '@/components/Web/Business/HoverIcon';
 import { IconType } from '@/components/Web/Business/HoverIcon/type';
 import ImageCrop, { ImageCropRef } from '@/components/Common/ImageCrop';
-import { cn, errorMessage } from '@/helper/utils';
+import { errorMessage } from '@/helper/ui';
+import { cn } from '@/helper/utils';
 import webApi from '@/service';
 import { Upload, message } from 'antd';
 import { RcFile } from 'antd/es/upload';
