+++ app/mobile/(profile)/user/profile/components/PersonalEdit/BasicInfoModal/index.tsx
@@ -1,7 +1,7 @@
 import Button from '@/components/Common/Button';
 import Input from '@/components/Common/Input';
 import Modal from '@/components/Common/Modal';
-import { errorMessage } from '@/helper/utils';
+import { errorMessage } from '@/helper/ui';
 import webApi from '@/service';
 import { useRequest } from 'ahooks';
 import { Form, message } from 'antd';
