+++ components/Web/Business/MiniElectiveDetailModal/index.tsx
@@ -1,6 +1,7 @@
 import Button from '@/components/Common/Button';
 import Modal from '@/components/Common/Modal';
-import { cn, errorMessage } from '@/helper/utils';
+import { errorMessage } from '@/helper/ui';
+import { cn } from '@/helper/utils';
 import Image from 'next/image';
 import { forwardRef, useImperativeHandle, useState } from 'react';
