+++ helper/utils.ts
@@ -1,7 +1,7 @@
+import * as React from 'react';
 import { Menu, QueryIdType } from '@/components/Web/Business/Breadcrumb/type';
 import { JumpLeaningLessonType } from '@/hooks/useCoursesHooks/useJumpLeaningLesson';
 import { CourseType } from '@/service/webApi/course/type';
-import { message } from 'antd';
 
 import { clsx, type ClassValue } from 'clsx';
 import { twMerge } from 'tailwind-merge';
@@ -105,13 +105,6 @@ export const deepClone = (obj: any) => {
   }
   return result;
 };
-
-export const errorMessage = (err: any) => {
-  const msg = err.msg || err.message;
-  msg && message.error(msg);
-};
-
-import * as React from 'react';
 
 export const { isValidElement } = React;
